dom/html/test/test_bug567938-1.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=567938
-->
<head>
  <title>Test for Bug 567938</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body onload="runTests();">
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=567938">Mozilla Bug 567938</a>
<p id="display"></p>
<iframe id='iframe' name="submit_frame" style="visibility: hidden;"></iframe>
<div id="content" style="display: none">
  <form id='f' method='get' target='submit_frame'>
  </form>
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 567938 **/

SimpleTest.waitForExplicitFinish();

var gTestData = ["submit", "image"];
var gCurrentTest = 0;

function initializeNextTest()
{
  var form = document.forms[0];

  // Cleaning-up.
  form.textContent = "";

  // Add the new element.
  var element = document.createElement("input");
  element.id = 'i';
  element.type = gTestData[gCurrentTest];
  element.onclick = function() { form.submit(); return false; };
  form.action = gTestData[gCurrentTest];
  form.appendChild(element);

  sendMouseEvent({type: 'click'}, 'i');
}

function runTests()
{
  document.getElementById('iframe').addEventListener('load', function(aEvent) {
    is(frames['submit_frame'].location.href,
       "http://mochi.test:8888/tests/dom/html/test/" + gTestData[gCurrentTest],
       "The form should have been submitted");
    gCurrentTest++;
    if (gCurrentTest < gTestData.length) {
      initializeNextTest();
    } else {
      aEvent.target.removeEventListener('load', arguments.callee);
      SimpleTest.finish();
    }
    });

  initializeNextTest();
}

</script>
</pre>
</body>
</html>
